buggy group chat
items are added to the database properly, but there are weird client side duplications.

When sending a message in group chat, it will show up twice but only for the sender client. Connection and disconnection notices are seen across all channels.

I have a feeling it has something to do with Broadcast vs Recieve from user, but i'm not super sure